add default connection factory
